+++ worker/crates/action-sink/src/file.rs
@@ -2,4 +2,5 @@ mod cesium3dtiles;
 mod excel;
 mod gltf;
 mod metadata;
+mod util;
 pub mod writer;
