+++ worker/crates/common/src/lib.rs
@@ -70,6 +70,7 @@ pub mod csv;
 pub mod dir;
 pub mod fs;
 pub mod future;
+pub mod gltf;
 pub mod json;
 pub mod serde;
 pub mod str;
